requirements.txt
aiohttp
akinator.py
APScheduler
bs4
discord.py
discord-pretty-help
discordpy-slash
pyparsing
psycopg2-binary
requests
sympy
TagScriptEngine